+++ als/julia/runner.jl
@@ -80,6 +80,7 @@ rm("als/data/users_t_test.dat")
 rm("als/data/users_t_train.dat")
 rm("als/data/movielens.cache")
 rm("als/data/movielens.reg")
+rm("als/data/movielens_preds.txt")
 testfile = open("test_results.txt", "a")
 write(testfile, "\nALS in JuliaVW\n")
 current_time = now()
