+++ criteo/README.md
@@ -1,3 +1,5 @@
 ## Using VP to predict Criteo ad clicks
 
 Using data from the [Criteo Kaggle Challenge](https://www.kaggle.com/c/criteo-display-ad-challenge), train on 45.8M rows of user ad clicks (or no clicks) and then test on 6M rows to generate predictions.
+
+Gets a log loss of 0.67848 in X min, which is enough to get position #612 on the leaderboard.
