+++ test_results.txt
@@ -1,36 +1,48 @@
 
 TITANIC IN R XGB
-2016-12-12 19:39:43
-AUC: 0.333333333333333
-Time: 1.61030697822571 sec
-Speed: 1805.27688141896 mcs/row
+2016-12-13 08:02:43
+AUC: 0.5
+Time: 1.53443098068237 sec
+Speed: 1720.21410390401 mcs/row
 
 TITANIC IN PYTHON XGB
-2016-12-12 19:39:44.785568
-AUC: 0.802816901408
-Time: 1.122275 sec
-Speed: 1258.1558296 mcs/row
+2016-12-13 08:02:44.809284
+AUC: 0.680250783699
+Time: 1.182247 sec
+Speed: 1325.38901345 mcs/row
 
 TITANIC IN PYTHON VP
-2016-12-12 19:39:45.296371
+2016-12-13 08:02:45.325962
 AUC: 0.847076923077
-Time: 0.467518 sec
-Speed: 524.123318386 mcs/row
+Time: 0.472407 sec
+Speed: 529.60426009 mcs/row
 
 TITANIC IN BASH VW
-Mon Dec 12 19:39:45 EST 2016
-AUC: 0.773148148148
-Time: 0.994 sec
-Speed: 1114.34977578475336322869 mcs/row
+Tue Dec 13 08:02:45 EST 2016
+AUC: 0.798575927701
+Time: 0.879 sec
+Speed: 985.42600896860986547085 mcs/row
 
 MOVIELENS ALS IN BASH VW
-Mon Dec 12 19:39:46 EST 2016
-RMSE: 0.913258
-Time: 38.788 sec
-Speed: 38.78800000000000000000 mcs/row
+Tue Dec 13 08:02:46 EST 2016
+RMSE: 0.924958
+Time: 34.493 sec
+Speed: 34.49300000000000000000 mcs/row
 
 ALS in JuliaVW
-2016-12-12T19:40:52.816
-RMSE: 0.9775263954118155
-Time: 27278 milliseconds
-Speed: 272.78000000000003 mcs/row
+2016-12-13T08:03:48.065
+RMSE: 0.963564945687125
+Time: 26935 milliseconds
+Speed: 269.34999999999997 mcs/row
+
+MOVIELENS IN PYSPARK
+2016-12-13 08:04:04.389238
+RMSE: 1.12096589905
+Time: 14.973381 sec
+Speed: 1497.3381 mcs per row
+
+MOVIELENS IN VP
+2016-12-13 08:06:35.462718
+RMSE: 0.994833783761
+Time: 131.723019 sec
+Speed: 131.695494642 mcs/row
